Refactor  some of our conversion code
We often have to do some slightly different conversions depending on if
we're making updates within just our app, or if we are receiving updates
from a caldav client. We want to put more of our conversion code in a
a single place to make it easier to understand in the future.